requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
pydantic~=2.10.6
starlette~=0.45.3
sqlalchemy~=2.0.37
passlib~=1.7.4
bcrypt~=4.0.1
python-multipart~=0.0.20
python-jose~=3.3.0
psycopg2-binary~=2.9.10
alembic~=1.14.1
pytest~=8.3.5
httpx~=0.28.1
pytest-asyncio~=0.25.3
aiofiles~=24.1.0
Jinja2~=3.1.6